move TestUtils to correct package
