+++ src/test/java/org/team199/wpiws/devices/DevicesTest.java
@@ -26,11 +26,11 @@ import org.mockito.ArgumentMatcher;
 import org.mockito.InOrder;
 import org.mockito.MockedStatic;
 import org.team199.wpiws.Pair;
+import org.team199.wpiws.TestUtils;
 import org.team199.wpiws.connection.ConnectionProcessor;
 import org.team199.wpiws.connection.WSValue;
 import org.team199.wpiws.interfaces.BooleanCallback;
 import org.team199.wpiws.interfaces.ObjectCallback;
-import org.team199.wpiws.interfaces.TestUtils;
 import org.team199.wpiws.interfaces.TriFunction;
 import org.team199.wpiws.types.LEDColor;
